RAG/requirements.txt
requests>=2.31.0
qdrant-client>=1.9.0
cohere>=4.4.3
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
trafilatura>=1.6.0